% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A1.R
\docType{data}
\name{A1}
\alias{A1}
\title{A trajectory of elephant sea}
\format{A trajectory  with 2218 rows and 7 columns:
\describe{
  \item{id}{Id each point of trajectory}
   \item{x}{Latitude}
  \item{y}{Longitude}
  \item{time}{Time stamps of the track points}
  \item{endtime}{End time stamps of the track points}
  \item{timeindex}{Time index}
  \item{ones}{Ones of the track} 
}}
\usage{
A1
}
\description{
object that contains distance, duration, speed and direction.
}
\keyword{datasets}
