% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A2.R
\docType{data}
\name{A2}
\alias{A2}
\title{A trajectory of elephant sea}
\format{A trajectory  with 2148 rows and 7 columns:
\describe{
  \item{id}{id each point of trajectory}
   \item{x}{Latitude}
  \item{y}{Longitude}
  \item{time}{Time stamps of the track points}
  \item{endtime}{End time stamps of the track points}
  \item{timeindex}{Id of time stamps}
  \item{ones}{Ones of the track}
   }}
\usage{
A2
}
\description{
A object that contains distance, duration, speed and direction.
}
\keyword{datasets}
