% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSpatialCluster.R
\docType{methods}
\name{createSpatialCluster}
\alias{createSpatialCluster}
\alias{createSpatialCluster,Track,list-method}
\title{Create Spatial Cluster}
\usage{
createSpatialCluster(A3, clusterlist)

\S4method{createSpatialCluster}{Track,list}(A3, clusterlist)
}
\arguments{
\item{A3}{Track object}

\item{clusterlist}{list of cluster positions create spatial objects}
}
\value{
cluster of polygons
}
\description{
Method for create a spatial cluster
}
\examples{

spcluster<-createSpatialCluster(A1,speedCluster(A1,mean(A1@connections$speed),
12,min(A1@connections$speed)+4))

}
\author{
Deigo Monteiro
}
