% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSourceInfo.R
\docType{class}
\name{DataSourceInfo-class}
\alias{DataSourceInfo-class}
\alias{DataSourceInfo}
\title{Data Source Info}
\description{
Class to connect in a database
}
\section{Slots}{

\describe{
\item{\code{user}}{User of database}

\item{\code{title}}{Title of database}

\item{\code{accessDriver}}{The database access driver}

\item{\code{host}}{Host of the database (e.g. localhost)}

\item{\code{port}}{Port the database (e.g. 5432)}

\item{\code{timeout}}{Timeout time of connection}

\item{\code{password}}{Password of database}

\item{\code{db}}{Database name}

\item{\code{encoding}}{Encoding of database (e.g.CP1252 )}

\item{\code{dbtype}}{The type of the database (e.g. POSTGIS)}

\item{\code{path}}{Path of the database}
}}

