% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DouglasPeucker.R
\docType{methods}
\name{douglasPeuckerRP}
\alias{douglasPeuckerRP}
\alias{douglasPeuckerRP,Track,numeric,numeric,numeric-method}
\title{Douglas Peucker RP}
\usage{
douglasPeuckerRP(A1, firstp, lastp, dist)

\S4method{douglasPeuckerRP}{Track,numeric,numeric,numeric}(A1, firstp, lastp,
  dist)
}
\arguments{
\item{A1}{Represents a single trajectory followed by a person, animal or object}

\item{firstp}{Given first point}

\item{lastp}{Given last point}

\item{dist}{Distance time series}
}
\description{
Method that reduces a trajectory spatially with first point and last point
}
\examples{

doug <- douglasPeuckerRP(A1,20,200,2000)

}
\author{
Diego Monteiro
}
