% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexToTrack.R
\name{IndexToTrack}
\alias{IndexToTrack}
\title{Convet index to Track}
\usage{
IndexToTrack(A1, index)
}
\arguments{
\item{A1}{Represents a single trajectory followed by a person, animal or object}

\item{index}{An index list}
}
\value{
Track object
}
\description{
Convet index to Track
}
\author{
Diego Monteiro
}
