% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LimitedNeighborhood.R
\name{LimitedNeighborhood}
\alias{LimitedNeighborhood}
\title{Limited Neighborhood}
\usage{
LimitedNeighborhood(track, ini, minT, cln, cl, avg, sl)
}
\arguments{
\item{track}{Track obeject}

\item{ini}{numeric parameter}

\item{minT}{is the minimun period at the speed}

\item{cln}{numeric value}

\item{cl}{numeric list}

\item{avg}{average value of speed}

\item{sl}{is the speed limit}
}
\description{
Check the limite of Neighborhood
}
