% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OWMeratniaBy.R
\docType{methods}
\name{owMeratniaBy}
\alias{owMeratniaBy}
\alias{owMeratniaBy,Track,numeric,numeric-method}
\title{Ow Meratinia By}
\usage{
owMeratniaBy(A1, dist, speed)

\S4method{owMeratniaBy}{Track,numeric,numeric}(A1, dist, speed)
}
\arguments{
\item{A1}{track object}

\item{dist}{distance of object}

\item{speed}{speed of object}
}
\value{
reduces trajectories spatiotemporally
}
\description{
Method that reduces trajectories spatiotemporally
}
