% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OWMeratniaByCollection.R
\name{owMeratniaByCollection}
\alias{owMeratniaByCollection}
\title{Ow Meratnial By Collection}
\usage{
owMeratniaByCollection(A1, dist, speed)
}
\arguments{
\item{A1}{TracksCollection object}

\item{dist}{Distance of object}

\item{speed}{Speed of object}
}
\value{
Trajectory spatiotemporally reduced
}
\description{
Method that reduces a set of trajectories spatiotemporally
}
