% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoAlong.R
\docType{methods}
\name{partner}
\alias{partner}
\alias{partner,Track,Track,numeric,numeric,numeric,DataSourceInfo,character-method}
\alias{partner,Track,Track,numeric,numeric,numeric,PostgreSQLConnection,character-method}
\alias{partner,Track,Track,numeric,numeric,numeric,logical,missing-method}
\alias{partner,TracksCollection,missing,numeric,numeric,numeric,missing,missing-method}
\alias{partner,TracksCollection,TracksCollection,numeric,numeric,numeric,missing,missing-method}
\alias{partner,TracksCollection,Track,numeric,numeric,numeric,missing,missing-method}
\title{Partner}
\usage{
partner(A1, A2, dist, maxtime, mintime, datasource, tablename)


  \S4method{partner}{Track,Track,numeric,numeric,numeric,DataSourceInfo,character}(A1,
  A2, dist, maxtime, mintime, datasource, tablename)


  \S4method{partner}{Track,Track,numeric,numeric,numeric,PostgreSQLConnection,character}(A1,
  A2, dist, maxtime, mintime, datasource, tablename)

\S4method{partner}{Track,Track,numeric,numeric,numeric,logical,missing}(A1, A2,
  dist, maxtime, mintime, datasource, tablename)


  \S4method{partner}{TracksCollection,missing,numeric,numeric,numeric,missing,missing}(A1,
  A2, dist, maxtime, mintime, datasource, tablename)


  \S4method{partner}{TracksCollection,TracksCollection,numeric,numeric,numeric,missing,missing}(A1,
  A2, dist, maxtime, mintime, datasource, tablename)


  \S4method{partner}{TracksCollection,Track,numeric,numeric,numeric,missing,missing}(A1,
  A2, dist, maxtime, mintime, datasource, tablename)
}
\arguments{
\item{A1}{represents a single trajectory followed by a person, animal or object.}

\item{A2}{represents a single trajectory followed by a person, animal or object.}

\item{dist}{distance time series}

\item{maxtime}{Is the maximum time of point object}

\item{mintime}{Is the minimum time of point object}

\item{datasource}{Is object class DataSourceInfo}

\item{tablename}{The name of the table database}
}
\value{
List with begin time and end time stamps of two  objects partner
}
\description{
Method to recognize trajectories that stay together, based on trajectory distance time series analysis
}
\examples{

partner(A1,A2,110792,2277,0,FALSE)

}
\author{
Diego Monteiro
}
