% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SendGoAlong.R
\docType{methods}
\name{sendPartnerPairsToDB}
\alias{sendPartnerPairsToDB}
\alias{sendPartnerPairsToDB,list,DataSourceInfo,character-method}
\alias{sendPartnerPairsToDB,list,PostgreSQLConnection,character-method}
\title{send Partner Pairs To DataBase}
\usage{
sendPartnerPairsToDB(dataframe, dataSourceInfo, tablename)

\S4method{sendPartnerPairsToDB}{list,DataSourceInfo,character}(dataframe,
  dataSourceInfo, tablename)

\S4method{sendPartnerPairsToDB}{list,PostgreSQLConnection,character}(dataframe,
  dataSourceInfo, tablename)
}
\arguments{
\item{dataframe}{dataframe list}

\item{dataSourceInfo}{a object class dataSourceInfo}

\item{tablename}{name of table}
}
\value{
send the partners list for a database
}
\description{
Method that sends found partners to a PostGIS database

Method that sends found partners to a PostGIS database with object DataSourceInfo

Method that sends found partners to a PostGIS database using object PosgreSQLConnection
}
