% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyCompareMethod.R
\docType{class}
\name{singledifftrack-class}
\alias{singledifftrack-class}
\title{Sigled diff track}
\arguments{
\item{tr1}{Track object}

\item{tr2}{Track object}
}
\value{
a difftrack object
}
\description{
My compare method get distances between 2 Track objects for each point in time where they overlap and create a corresponding line
}
