% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlowestNeighborhood.R
\name{SlowestNeighborhood}
\alias{SlowestNeighborhood}
\title{Slowest Neighborhood}
\usage{
SlowestNeighborhood(track, ini, minT, cl)
}
\arguments{
\item{track}{Track obeject}

\item{ini}{numeric parameter}

\item{minT}{is the minimun period at the speed}

\item{cl}{numeric list}
}
\description{
Method for check slowest neighborhood
}
