% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeedCluster.R
\name{speedCluster}
\alias{speedCluster}
\title{Speed Cluster}
\usage{
speedCluster(track, avg, minT, sl)
}
\arguments{
\item{track}{Track object}

\item{avg}{is the average speed}

\item{minT}{is the minimun period at the speed}

\item{sl}{is the speed limit}
}
\value{
returns regions where speed was lower than the defined parameter
}
\description{
Method for check the regions where speed was lower than the defined parameter
}
\details{
Order the speed so it will start with the slowest speed cluster
}
