% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracksCollection.R
\docType{data}
\name{tracksCollection}
\alias{tracksCollection}
\title{A tracks colleection of a plataform argos}
\format{A trajectory  with 7 columns:
\describe{
  \item{id}{Id rows}
   \item{xmin}{Latitude minimum}
   \item{xmax}{Latitude maximum}
  \item{ymin}{Longitude minimum}
  \item{ymax}{Longitude maximum}
  \item{tmax}{Maximum time}
  \item{tmin}{Minimum time}
  \item{timeindex}{Time index}
  \item{n}{Ones of the track} 
}}
\usage{
tracksCollection
}
\description{
A dataset containing tracks objects
}
\keyword{datasets}
