% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A1.R
\docType{data}
\name{A1}
\alias{A1}
\title{A trajectory of elephant sea}
\format{A trajectory  with 2218 rows and 7 columns:
\describe{
  \item{id}{id rows}
   \item{x}{latitude}
  \item{y}{longitude}]
  \item{time}{begin time}
  \item{endtime}{end time}
  \item{timeindex}{time index}
  \item{ones}{ones of the track} 
}}
\usage{
A1
}
\description{
A dataset containing the track object
}
\keyword{datasets}
