% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A1.R
\docType{data}
\name{A1}
\alias{A1}
\title{A trajectory of elephant sea}
\format{A trajectory  with 2218 rows and 7 columns:
\describe{
  \item{id}{id each point of trajectory}
   \item{x}{latitude}
  \item{y}{longitude}]
  \item{time}{time stamps of the track points}
  \item{endtime}{end time stamps of the track points}
  \item{timeindex}{time index}
  \item{ones}{ones of the track} 
}}
\usage{
A1
}
\description{
object that contains distance, duration, speed and direction.
}
\keyword{datasets}
