% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A2.R
\docType{data}
\name{A2}
\alias{A2}
\title{A trajectory of elephant sea}
\format{A trajectory  with 2148 rows and 7 columns:
\describe{
  \item{id}{id each point of trajectory}
   \item{x}{latitude}
  \item{y}{longitude}
  \item{time}{time stamps of the track points}
  \item{endtime}{end time stamps of the track points}
  \item{timeindex}{id of time stamps}
  \item{ones}{ones of the track}
   }}
\usage{
A2
}
\description{
A object that contains distance, duration, speed and direction.
}
\keyword{datasets}
