% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A2.R
\docType{data}
\name{A2}
\alias{A2}
\title{A trajectory of elephant sea}
\format{A trajectory  with 2148 rows and 7 columns:
\describe{
  \item{id}{id rows}
   \item{x}{latitude}
  \item{y}{longitude}]
  \item{time}{begin time}
  \item{endtime}{end time}
  \item{timeindex}{time index}
  \item{ones}{ones of the track} 
}}
\usage{
A2
}
\description{
A dataset containing the track obejct
}
\keyword{datasets}
