% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSpatialCluster.R
\name{createSpatialCluster}
\alias{createSpatialCluster}
\title{Create Spatial Cluster}
\usage{
createSpatialCluster(A3, clusterlist)
}
\arguments{
\item{clusterlist}{list of cluster positions create spatial objects}

\item{track}{Track object}
}
\value{
Polygons
}
\description{
Method for create a spatial cluster
}
