% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSourceInfo.R
\docType{class}
\name{DataSourceInfo-class}
\alias{DataSourceInfo-class}
\alias{DataSourceInfo}
\title{Data Source Info}
\description{
Class to connect in a database
}
\section{Slots}{

\describe{
\item{\code{user}}{user of database}

\item{\code{title}}{Title of database}

\item{\code{accessDriver}}{the database access driver}

\item{\code{host}}{host of the database (e.g. local host)}

\item{\code{port}}{port the database (e.g. 5432)}

\item{\code{timeout}}{timeout time of connection}

\item{\code{password}}{password of database}

\item{\code{db}}{database name}

\item{\code{encoding}}{encoding of database (e.g.CP1252 )}

\item{\code{dbtype}}{the type of the database (e.g. POSTGIS)}

\item{\code{path}}{path of the database}
}}

