% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSourceInfo.R
\docType{class}
\name{DataSourceInfo-class}
\alias{DataSourceInfo-class}
\alias{DataSourceInfo}
\title{Data Source Info}
\description{
Class to connect in a database
}
\section{Slots}{

\describe{
\item{\code{user}}{user of database}

\item{\code{password}}{password of database}

\item{\code{db}}{database name}
}}

