% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectionCluster.R
\docType{methods}
\name{directionCluster}
\alias{directionCluster}
\alias{directionCluster,Track,numeric,numeric,numeric-method}
\title{Direction Cluster}
\usage{
directionCluster(track, minD, minT, tolerance)

\S4method{directionCluster}{Track,numeric,numeric,numeric}(track, minD, minT,
  tolerance)
}
\arguments{
\item{track}{Track object}

\item{minD}{is the minimun direction change}

\item{minT}{is the minimun period at the speed}

\item{tolerance}{is the maximum change parameter}
}
\value{
returns regions where direction changed more than the defined parameter
}
\description{
Direction Cluster
}
