% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DouglasPeucker.R
\name{douglasPeucker}
\alias{douglasPeucker}
\title{Douglas Peucker}
\usage{
douglasPeucker(A1, dist)
}
\arguments{
\item{A1}{track object trajectories}

\item{dist}{distance}
}
\value{
track reduced  trajectory spatially
}
\description{
Method that reduces a trajectory spatially
}
\examples{

douglasPeucker(A1,300)


}
