% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DouglasPeucker.R
\docType{methods}
\name{douglasPeuckerRP,Track,numeric,numeric,numeric-method}
\alias{douglasPeuckerRP,Track,numeric,numeric,numeric-method}
\title{Douglas Peucker RP}
\usage{
\S4method{douglasPeuckerRP}{Track,numeric,numeric,numeric}(A1, firstp, lastp,
  dist)
}
\arguments{
\item{A1}{track object}

\item{firstp}{first point}

\item{lastp}{last point}

\item{dist}{distance of object}
}
\description{
Douglas Peucker RP
}
