% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DouglasPeucker.R
\docType{methods}
\name{douglasPeuckerRP}
\alias{douglasPeuckerRP}
\alias{douglasPeuckerRP,Track,numeric,numeric,numeric-method}
\title{Douglas Peucker RP}
\usage{
douglasPeuckerRP(A1, firstp, lastp, dist)

\S4method{douglasPeuckerRP}{Track,numeric,numeric,numeric}(A1, firstp, lastp,
  dist)
}
\arguments{
\item{A1}{track object}

\item{firstp}{given first point}

\item{lastp}{given last point}

\item{dist}{distance}
}
\description{
Method that reduces a trajectory spatially with first point and last point
}
\examples{

doug <- douglasPeuckerRP(A1,20,200,2000)

}
\author{
Diego Monteiro
}
