% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexToTrack.R
\name{IndexToTrack}
\alias{IndexToTrack}
\title{Conversão de index pra track}
\usage{
IndexToTrack(A1, index)
}
\arguments{
\item{A1}{Track object}

\item{index}{an index list}
}
\value{
track object
}
\description{
Conversão de index pra track
}
