% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LimitedNeighborhood.R
\name{LimitedNeighborhood}
\alias{LimitedNeighborhood}
\title{Limited Neighborhood}
\usage{
LimitedNeighborhood(track, ini, minT, cln, cl, avg, sl)
}
\arguments{
\item{track}{Represents a single trajectory followed by a person, animal or object}

\item{ini}{Order list of track speed}

\item{minT}{Is the minimun period at the speed}

\item{cln}{Cluster identification}

\item{cl}{Empty list}

\item{avg}{Average value of speed}

\item{sl}{Is the speed limit}
}
\description{
Check the limite of Neighborhood
}
\author{
Diego Monteiro
}
