% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OWMeratniaByCollection.R
\docType{methods}
\name{owMeratniaByCollection}
\alias{owMeratniaByCollection}
\alias{owMeratniaByCollection,TracksCollection,numeric,numeric-method}
\title{Ow Meratnial By Collection}
\usage{
owMeratniaByCollection(A1, dist, speed)

\S4method{owMeratniaByCollection}{TracksCollection,numeric,numeric}(A1, dist,
  speed)
}
\arguments{
\item{A1}{represents a collection of trajectories followed by different persons, animals or objects}

\item{dist}{Distance of object}

\item{speed}{Speed of object}
}
\value{
Trajectory spatiotemporally reduced
}
\description{
Method that reduces a set of trajectories spatiotemporally
}
\examples{

library(magrittr)

library(ggplot2)
ow <-owMeratniaByCollection(tracksCollection,13804.84 ,0.03182201) \%>\% coordinates()

df <- data.frame(x=ow[,1],y=ow[,2])

ggplot(df,aes(x=x,y=y))+geom_path(aes(group = 1), arrow = arrow(),color='blue')
}
\author{
Diego Monteiro
}
