% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OWMeratniaByCollection.R
\docType{methods}
\name{owMeratniaByCollection}
\alias{owMeratniaByCollection}
\alias{owMeratniaByCollection,TracksCollection,numeric,numeric-method}
\title{Ow Meratnial By Collection}
\usage{
owMeratniaByCollection(A1, dist, speed)

\S4method{owMeratniaByCollection}{TracksCollection,numeric,numeric}(A1, dist,
  speed)
}
\arguments{
\item{A1}{TracksCollection object}

\item{dist}{Distance of object}

\item{speed}{Speed of object}
}
\value{
Trajectory spatiotemporally reduced
}
\description{
Method that reduces a set of trajectories spatiotemporally
}
