% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoAlong.R
\name{partner}
\alias{partner}
\title{Partner}
\usage{
partner(A1, A2, dist, maxtime, mintime, datasource, tablename)
}
\arguments{
\item{A1}{Track object}

\item{A2}{Track object}

\item{dist}{distance of object}

\item{maxtime}{the maximum time}

\item{mintime}{the minimum time}

\item{datasource}{object class DataSourceInfo}

\item{tablename}{The table name}
}
\value{
list with begin time and end time of two  objects partner
}
\description{
Method to identify if two trajectories are partners
}
