% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoAlong.R
\docType{methods}
\name{partner}
\alias{partner}
\alias{partner,Track,Track,numeric,numeric,numeric,DataSourceInfo,character-method}
\alias{partner,Track,Track,numeric,numeric,numeric,PostgreSQLConnection,character-method}
\alias{partner,Track,Track,numeric,numeric,numeric,logical,missing-method}
\alias{partner,TracksCollection,missing,numeric,numeric,numeric,missing,missing-method}
\alias{partner,TracksCollection,TracksCollection,numeric,numeric,numeric,missing,missing-method}
\alias{partner,TracksCollection,Track,numeric,numeric,numeric,missing,missing-method}
\title{Partner}
\usage{
partner(A1, A2, dist, maxtime, mintime, datasource, tablename)


  \S4method{partner}{Track,Track,numeric,numeric,numeric,DataSourceInfo,character}(A1,
  A2, dist, maxtime, mintime, datasource, tablename)


  \S4method{partner}{Track,Track,numeric,numeric,numeric,PostgreSQLConnection,character}(A1,
  A2, dist, maxtime, mintime, datasource, tablename)

\S4method{partner}{Track,Track,numeric,numeric,numeric,logical,missing}(A1, A2,
  dist, maxtime, mintime, datasource, tablename)


  \S4method{partner}{TracksCollection,missing,numeric,numeric,numeric,missing,missing}(A1,
  A2, dist, maxtime, mintime, datasource, tablename)


  \S4method{partner}{TracksCollection,TracksCollection,numeric,numeric,numeric,missing,missing}(A1,
  A2, dist, maxtime, mintime, datasource, tablename)


  \S4method{partner}{TracksCollection,Track,numeric,numeric,numeric,missing,missing}(A1,
  A2, dist, maxtime, mintime, datasource, tablename)
}
\arguments{
\item{A1}{Track object}

\item{A2}{Track object}

\item{dist}{distance of object}

\item{maxtime}{the maximum time}

\item{mintime}{the minimum time}

\item{datasource}{object class DataSourceInfo}

\item{tablename}{The table name}
}
\value{
list with begin time and end time of two  objects partner
}
\description{
Method to identify if two trajectories are partners

DataSourceInfo connection

Postgresql connection

Logical param in database connection

Missing database connection

Missing database connection and TrackCollection objects

Missing database connection
}
