% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SizeVerifier.R
\name{RightSize}
\alias{RightSize}
\title{Right size verifier}
\usage{
RightSize(diffTracks, begintime, endtime, sizeMultiplier)
}
\arguments{
\item{diffTracks}{difftrack}

\item{begintime}{begin time}

\item{endtime}{end time}

\item{sizeMultiplier}{number will multiplier the diff}
}
\value{
boolean
}
\description{
Right size verifier
}
