% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SendGoAlong.R
\name{sendPartnerPairsToDB}
\alias{sendPartnerPairsToDB}
\title{send Partner Pairs To DataBase}
\usage{
sendPartnerPairsToDB(dataframe, dataSourceInfo, tablename)
}
\arguments{
\item{dataframe}{dataframe list}

\item{dataSourceInfo}{a object class dataSourceInfo}

\item{tablename}{name of table}
}
\value{
send the partners list for a database
}
\description{
Method that sends found partners to a PostGIS database
}
