% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyCompareMethod.R
\docType{class}
\name{singledifftrack-class}
\alias{singledifftrack-class}
\title{Sigle diff track}
\arguments{
\item{tr1}{represents a single trajectory followed by a person, animal or object.}

\item{tr2}{represents a single trajectory followed by a person, animal or object.}
}
\value{
a difftrack object
}
\description{
My compare method get distances between 2 Track objects for each point in time where they overlap and create a corresponding line
}
\details{
@import xts
}
\author{
Diego Monteiro
}
