% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlowestNeighborhood.R
\name{SlowestNeighborhood}
\alias{SlowestNeighborhood}
\title{Slowest Neighborhood}
\usage{
SlowestNeighborhood(track, ini, minT, cl)
}
\arguments{
\item{track}{Represents a single trajectory followed by a person, animal or object}

\item{ini}{Order list of track speed}

\item{minT}{Is the minimun period at the speed}

\item{cl}{Empty list}
}
\description{
Method for check slowest neighborhood
}
\author{
Diego Monteiro
}
