% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeedCluster.R, R/SpeedFilter.R
\docType{methods}
\name{speedCluster,Track,numeric,numeric,numeric-method}
\alias{speedCluster,Track,numeric,numeric,numeric-method}
\alias{speedFilter}
\alias{speedFilter,Track,numeric-method}
\title{Speed filter}
\usage{
\S4method{speedCluster}{Track,numeric,numeric,numeric}(track, avg, minT, sl)

speedFilter(A1, speed)

\S4method{speedFilter}{Track,numeric}(A1, speed)
}
\arguments{
\item{A1}{Track object}

\item{speed}{is the maximum speed parameter}
}
\description{
Method to return where speed was higher than the defined parameter
}
\examples{
library(ggplot2)

speed <- min(A1@connections$speed)

sf <- speedFilter(A1,speed)

df <- data.frame(x=sf@sp@coords[,1],y=sf@sp@coords[,2])

ggplot(df,aes(x=df$x,y=df$y))+geom_path(aes(group = 1), arrow = arrow(),color='blue')

}
\author{
Diego Monteiro
}
