% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeedFilter.R
\name{speedFilter}
\alias{speedFilter}
\title{Speed filter}
\usage{
speedFilter(A1, speed)
}
\arguments{
\item{A1}{Track object}

\item{speed}{is the maximum speed parameter}
}
\description{
Method to return where speed was higher than the defined parameter
}
