% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracksCollection.R
\docType{data}
\name{tracksCollection}
\alias{tracksCollection}
\title{A tracks colleection of a plataform argos}
\format{A trajectory  with 7 columns:
\describe{
  \item{id}{id rows}
   \item{xmin}{latitude minimum}
   \item{xmax}{latitude maximum}
  \item{ymin}{longitude minimum}
  \item{ymax}{longitude maximum}
  \item{tmax}{maximum time}
  \item{tmin}{minimum time}
  \item{timeindex}{time index}
  \item{n}{ones of the track} 
}}
\usage{
tracksCollection
}
\description{
A dataset containing tracks objects
}
\keyword{datasets}
